tools/network_analyzer.py
import asyncio
from scapy.all import *
import nmap
from typing import Dict, List, Optional
import logging
from concurrent.futures import ThreadPoolExecutor
import json

logger = logging.getLogger(__name__)

class AdvancedNetworkAnalyzer:
    def __init__(self, timeout: int = 120, buffer_size: int = 8192, analysis_depth: int = 3):
        self.timeout = timeout
        self.buffer_size = buffer_size
        self.analysis_depth = analysis_depth
        self.nmap_scanner = nmap.PortScanner()
        self._executor = ThreadPoolExecutor(max_workers=4)
    
    async def deep_scan(self, target: str) -> Dict:
        """Realiza un análisis profundo de la red."""
        try:
            tasks = [
                self._port_scan(target),
                self._vulnerability_scan(target),
                self._traffic_analysis(target)
            ]
            results = await asyncio.gather(*tasks)
            return self._consolidate_results(results)
        except Exception as e:
            logger.error(f"Error en análisis profundo: {e}")
            raise

    async def _port_scan(self, target: str) -> Dict:
        """Escaneo de puertos asíncrono."""
        def _scan():
            return self.nmap_scanner.scan(target, arguments="-sS -sV -O")
        return await asyncio.get_event_loop().run_in_executor(self._executor, _scan)

    async def _vulnerability_scan(self, target: str) -> Dict:
        """Análisis de vulnerabilidades."""
        def _scan():
            return self.nmap_scanner.scan(target, arguments="-sC --script vuln")
        return await asyncio.get_event_loop().run_in_executor(self._executor, _scan)

    async def _traffic_analysis(self, target: str) -> Dict:
        """Análisis de tráfico de red."""
        packets = []
        def packet_callback(pkt):
            if len(packets) < self.buffer_size:
                packets.append(pkt.summary())
        
        try:
            sniff(filter=f"host {target}", prn=packet_callback, 
                  timeout=self.timeout, store=0)
            return {"packets": packets}
        except Exception as e:
            logger.error(f"Error en análisis de tráfico: {e}")
            return {"error": str(e)}

    def _consolidate_results(self, results: List[Dict]) -> Dict:
        """Consolida los resultados de diferentes análisis."""
        return {
            "port_scan": results[0],
            "vulnerabilities": results[1],
            "traffic_analysis": results[2],
            "analysis_timestamp": datetime.now().isoformat()
        }

class ParallelProcessor:
    def __init__(self, batch_size: int = 64):
        self.batch_size = batch_size
        self._executor = ThreadPoolExecutor(max_workers=8)
        
    async def process_batch(self, items: List, processor_func) -> List:
        """Procesa un lote de items en paralelo."""
        batches = [items[i:i + self.batch_size] 
                  for i in range(0, len(items), self.batch_size)]
        
        async def process_single_batch(batch):
            tasks = [self._process_item(item, processor_func) for item in batch]
            return await asyncio.gather(*tasks)
        
        results = []
        for batch in batches:
            batch_results = await process_single_batch(batch)
            results.extend(batch_results)
        return results
    
    async def _process_item(self, item, processor_func):
        """Procesa un único item."""
        try:
            return await asyncio.get_event_loop().run_in_executor(
                self._executor, processor_func, item
            )
        except Exception as e:
            logger.error(f"Error procesando item: {e}")
            return {"error": str(e), "item": item}
