requirements.txt
# AI y Machine Learning
anthropic==0.8.0
openai==1.12.0
langchain==0.1.0
transformers==4.36.0
torch==2.1.0
tensorflow==2.15.0

# Análisis de Datos y Finanzas
pandas==2.2.1
numpy==1.26.4
scipy==1.11.0
yfinance==0.2.35
python-dotenv==1.0.0

# Seguridad y Criptografía
cryptography==42.0.4
pyOpenSSL==23.0.0
scapy==2.5.0
pycryptodome==3.19.0
python-nmap==0.7.1

# Web y APIs
fastapi==0.104.0
uvicorn==0.24.0
requests==2.31.0
aiohttp==3.9.0

# Interfaz de Usuario
streamlit==1.32.0
plotly==5.19.0
dash==2.14.0
gradio==4.0.0

# Testing y Desarrollo
pytest==7.4.0
black==23.11.0
mypy==1.7.0
pytest-asyncio==0.23.5
pytest==8.0.0
pytest-asyncio==0.23.5
aiohttp==3.9.3
cryptography==42.0.4
black==24.1.1
mypy==1.8.0
pylint==3.0.3
requests==2.31.0
cryptography==42.0.4
numpy==1.26.4
pandas==2.2.1
python-jose==3.3.0
python-dateutil==2.8.2
pycryptodome==3.20.0
tiktoken==0.6.0
